+++ source/extensions/omni.isaac.core/omni/isaac/core/tests/__init__.py
@@ -3,5 +3,6 @@ from .test_semantics import *
 from .test_xform_prim_pose import *
 from .prims.test_rigid_prim import *
 from .utils.test_rotations import *
+from .utils.test_stage import *
 from .world.test_world import *
 from .simulation_context.test_singleton import *